feat: 온보딩시 대피소 정보 로컬 저장
